在对外接口头文件中定义错误码



# message auto-generated for no-merge-commit merge:
!363 merge develop_error_code into develop

在对外接口头文件中定义错误码

Created-by: liu1103xwxw
Commit-by: l00915220
Merged-by: chenyz6
Description: # Pull Request Template

## Description

在对外接口头文件中定义错误码

## Related Issues

<!-- Replace with actual issue numbers. Example: Closes #123, Related to #456 -->

## Type of Change

- [ ] Bug fix
- [ ] New feature
- [ ] Performance optimization
- [ ] Documentation update
- [ ] Other (please describe):_______________

## Testing

<!-- Describe how the changes were tested. Include:
- Unit/integration test coverage
- Manual validation steps
- Links to CI reports (e.g., CIDA) if available
- Note any pending tests
-->

## Checklist

- [ ] I have performed a self-review of my own code.  
- [ ] I have updated the documentation.  
- [ ] I have added tests to prove my changes are effective.

See merge request: Ascend/memfabric_hybrid!363

+++ src/hybm/csrc/mm/hybm_dev_legacy_segment.cpp
@@ -298,7 +298,7 @@ Result HybmDevLegacySegment::GetExportSliceSize(size_t &size) noexcept
 // import可重入
 Result HybmDevLegacySegment::Import(const std::vector<std::string> &allExInfo, void *addresses[]) noexcept
 {
-    std::map<uint16_t, HbmExportInfo> importMap;
+    std::map<uint32_t, HbmExportInfo> importMap;
     LiteralExInfoTranslater<HbmExportInfo> translator;
     std::vector<HbmExportInfo> desInfos{};
     uint32_t localIdx = UINT32_MAX;
