+++ src/smem/include/host/smem_bm_def.h
@@ -89,7 +89,7 @@ typedef struct {
     bool dynamicWorldSize;            /* member cannot join dynamically */
     bool unifiedAddressSpace;         /* unified address with SVM */
     bool autoRanking;                 /* automatically allocate rank IDs, default is false. */
-    uint16_t rankId;                  /* user specified rank ID, valid for autoRanking is False */
+    uint32_t rankId;                  /* user specified rank ID, valid for autoRanking is False */
     uint32_t flags;                   /* other flag, default 0 */
     char hcomUrl[64];
     smem_tls_config hcomTlsConfig;
